+++ src/main/java/io/contek/invoker/deribit/api/common/_Stats.java
@@ -1,8 +1,10 @@
 package io.contek.invoker.deribit.api.common;
 
+import java.math.BigDecimal;
+
 public class _Stats {
   public Double high;
   public Double low;
   public Double price_change;
-  public Double volume;
+  public BigDecimal volume;
 }
